+++ service/controller/resource/resource.go
@@ -13,8 +13,9 @@ import (
 	"github.com/giantswarm/prometheus-meta-operator/service/controller/resource/alerting/alertmanagerconfig"
 	"github.com/giantswarm/prometheus-meta-operator/service/controller/resource/alerting/heartbeat"
 	"github.com/giantswarm/prometheus-meta-operator/service/controller/resource/alerting/heartbeatrouting"
+	alertingingress "github.com/giantswarm/prometheus-meta-operator/service/controller/resource/alerting/ingress"
 	"github.com/giantswarm/prometheus-meta-operator/service/controller/resource/certificates"
-	"github.com/giantswarm/prometheus-meta-operator/service/controller/resource/monitoring/ingress"
+	monitoringingress "github.com/giantswarm/prometheus-meta-operator/service/controller/resource/monitoring/ingress"
 	"github.com/giantswarm/prometheus-meta-operator/service/controller/resource/monitoring/prometheus"
 	"github.com/giantswarm/prometheus-meta-operator/service/controller/resource/monitoring/remotewriteconfig"
 	"github.com/giantswarm/prometheus-meta-operator/service/controller/resource/monitoring/scrapeconfigs"
@@ -42,7 +43,8 @@ type Config struct {
 	Region       string
 	Registry     string
 
-	OpsgenieKey string
+	AlertmanagerBaseDomain string
+	OpsgenieKey            string
 
 	PrometheusAddress             string
 	PrometheusBaseDomain          string
@@ -101,14 +103,14 @@ func New(config Config) ([]resource.Interface, error) {
 		}
 	}
 
-	var alertmanagerConfig resource.Interface
+	var alertmanagerConfigResource resource.Interface
 	{
 		c := alertmanagerconfig.Config{
 			K8sClient: config.K8sClient,
 			Logger:    config.Logger,
 		}
 
-		alertmanagerConfig, err = alertmanagerconfig.New(c)
+		alertmanagerConfigResource, err = alertmanagerconfig.New(c)
 		if err != nil {
 			return nil, microerror.Mask(err)
 		}
@@ -189,9 +191,25 @@ func New(config Config) ([]resource.Interface, error) {
 		}
 	}
 
-	var ingressResource resource.Interface
+	var alertingIngressResource resource.Interface
 	{
-		c := ingress.Config{
+		c := alertingingress.Config{
+			K8sClient:               config.K8sClient,
+			Logger:                  config.Logger,
+			BaseDomain:              config.AlertmanagerBaseDomain,
+			RestrictedAccessEnabled: config.RestrictedAccessEnabled,
+			WhitelistedSubnets:      config.WhitelistedSubnets,
+		}
+
+		alertingIngressResource, err = alertingingress.New(c)
+		if err != nil {
+			return nil, microerror.Mask(err)
+		}
+	}
+
+	var monitoringIngressResource resource.Interface
+	{
+		c := monitoringingress.Config{
 			K8sClient:               config.K8sClient,
 			Logger:                  config.Logger,
 			BaseDomain:              config.PrometheusBaseDomain,
@@ -199,7 +217,7 @@ func New(config Config) ([]resource.Interface, error) {
 			WhitelistedSubnets:      config.WhitelistedSubnets,
 		}
 
-		ingressResource, err = ingress.New(c)
+		monitoringIngressResource, err = monitoringingress.New(c)
 		if err != nil {
 			return nil, microerror.Mask(err)
 		}
@@ -238,12 +256,13 @@ func New(config Config) ([]resource.Interface, error) {
 	resources := []resource.Interface{
 		namespaceResource,
 		apiCertificatesResource,
-		alertmanagerConfig,
+		alertmanagerConfigResource,
 		scrapeConfigResource,
 		remoteWriteConfigResource,
 		prometheusResource,
 		verticalPodAutoScalerResource,
-		ingressResource,
+		alertingIngressResource,
+		monitoringIngressResource,
 		heartbeatResource,
 		heartbeatRoutingResource,
 	}